Updated migrations for price history monthly/weekly and dividend history, added next dividend information into dividend history fetch call, updated volume to bigint

+++ db/migrate/20220228040726_create_price_history_monthlies.rb
@@ -1,15 +1,14 @@
 class CreatePriceHistoryMonthlies < ActiveRecord::Migration[7.0]
   def change
     create_table :price_history_monthlies do |t|
-      t.integer :open, null: false
-      t.integer :low, null: false
-      t.integer :high, null: false
-      t.integer :close, null: false
-      t.integer :adjusted_close, null: false
-      t.integer :volume, null: false
+      t.float :open, null: false
+      t.float :low, null: false
+      t.float :high, null: false
+      t.float :close, null: false
+      t.float :adjusted_close, null: false
+      t.bigint :volume, null: false
       t.datetime :date, null: false
       t.references :company, null: false, foreign_key: true
-
       t.timestamps
     end
   end
